\name{assert.equal}
\alias{assert.equal}
\title{Assert equal}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Assert expected and actual are equal
}
\usage{assert.equal(expected, actual, msg = "")}
\arguments{
\item{expected}{}
\item{actual}{}
\item{msg}{}
}

\details{}

\examples{}
\keyword{debugging}
