\name{assert.isS3class}
\alias{assert.isS3class}
\title{Assert is S3 class}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Assert argument is of given S3 class
}
\usage{assert.isS3class(obj, class, msg = "")}
\arguments{
\item{obj}{}
\item{class}{}
\item{msg}{}
}

\details{}

\examples{}
\keyword{debugging}
