\name{assert}
\alias{assert}
\title{Assert}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Basic assert function that powers all others.
}
\usage{assert(condition, msg = "")}
\arguments{
\item{condition}{condition to evaluation for truth}
\item{msg}{message to display if condition is false}
\item{}{should warnings be displayed}
}

\details{}

\examples{}
\keyword{debugging}
