\name{assert.warning}
\alias{assert.warning}
\title{Assert warning}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Assert that a warning occured
}
\usage{assert.warning(f, warning = "", msg = "")}
\arguments{
\item{f}{}
\item{warning}{}
\item{msg}{}
}

\details{}

\examples{}
\keyword{debugging}
