\name{assert.warning_error}
\alias{assert.warning_error}
\title{Assert warning or error}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Assert a warning of error occured
}
\usage{assert.warning_error(w, warning = "", msg = "")}
\arguments{
\item{w}{}
\item{warning}{}
\item{msg}{}
}

\details{}

\examples{}
\keyword{debugging}
