\name{benchmark2}
\alias{benchmark2}
\title{Benchmark2}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
An extension to benchmark which allows you to specify a (minimum) time for the functions to run
}
\usage{benchmark2(..., time = 3, reps = 10)}
\arguments{
\item{...}{functions to compare, make sure to name them (see example)}
\item{time}{(minimum) time, in seconds, to run for}
\item{reps}{number of reps to run between checking time}
}

\details{}

\examples{}
\keyword{debugging}
