\name{displayCallNode}
\alias{displayCallNode}
\title{Display call node}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
@keyword debugging
}
\usage{displayCallNode(calls, level, value, depth, mintime)}
\arguments{
\item{calls}{}
\item{level}{}
\item{value}{}
\item{depth}{}
\item{mintime}{}
}

\details{}

\examples{}
\keyword{debugging}
