\name{expression.test}
\alias{expression.test}
\title{Expression test.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Run test and display status
}
\usage{expression.test(expression, print=TRUE)}
\arguments{
\item{expression}{expression to test}
\item{print}{print results?}
}

\details{}

\examples{}
\keyword{debugging}
