\name{getCalls}
\alias{getCalls}
\title{Get calls}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get all calls at a given level of the call stack
}
\usage{getCalls(calls, level, value)}
\arguments{
\item{calls}{list of calls}
\item{level}{level of call stack}
\item{value}{function name}
}

\details{}

\examples{}
\keyword{debugging}
