\name{is.positive.integer}
\alias{is.positive.integer}
\title{is.positive.integer}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Converience function for determining if x is a positive integer
}
\usage{is.positive.integer(x)}
\arguments{
\item{x}{object to test}
}

\details{}

\examples{}
\keyword{debugging}
