\name{is.test}
\alias{is.test}
\title{Is this expression a test?}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Determine whether expression is a test and should be run when "test"-ing a file
}
\usage{is.test(expression)}
\arguments{
\item{expression}{expression to test}
}

\details{}

\examples{}
\keyword{debugging}
