\name{plot.call.tree}
\alias{plot.call.tree}
\title{Plot call tree}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Attractively plot call tree
}
\usage{plot.call.tree(x, startlevel = 1, depth = 5, mintime = 2, ...)}
\arguments{
\item{x}{}
\item{startlevel}{}
\item{depth}{}
\item{mintime}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{debugging}
