\name{plotCallNode}
\alias{plotCallNode}
\title{Plot call node}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Recursively function that powers plot.call.tree
}
\usage{plotCallNode(calls, level, value, start, end, mintime, startlevel, depth)}
\arguments{
\item{calls}{}
\item{level}{}
\item{value}{}
\item{start}{}
\item{end}{}
\item{mintime}{}
\item{startlevel}{}
\item{depth}{}
}

\details{}

\examples{}
\keyword{debugging}
