\name{print.benchmark}
\alias{print.benchmark}
\title{Print benchmark}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print nicely formatted benchmark results
}
\usage{print.benchmark(x, type="total", ...)}
\arguments{
\item{x}{benchmark object to display}
\item{type}{time to display (one of system, user or total)}
\item{...}{required to match generic}
}

\details{}

\examples{}
\keyword{debugging}
