\name{print.call.tree}
\alias{print.call.tree}
\title{Print call tree}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Attractively print call tree
}
\usage{print.call.tree(x, startlevel = 3, depth = 8, mintime = 2, ...)}
\arguments{
\item{x}{}
\item{startlevel}{}
\item{depth}{}
\item{mintime}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{debugging}
