\name{test}
\alias{test}
\title{Test.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Run all tests in given file and return results
}
\usage{test(path, print=TRUE)}
\arguments{
\item{path}{path to file}
\item{print}{}
}

\details{Displays status for each test as run and then reports particular errors
...F..E.. 8/10 successful, then report errors}

\examples{}
\keyword{debugging}
