\name{test.result}
\alias{test.result}
\title{Eval expression and return list with errors, warnings and the deparsed expression}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
@keyword debugging
}
\usage{test.result(expression)}
\arguments{
\item{expression}{}
}

\details{}

\examples{}
\keyword{debugging}
