\name{throw_assert_error}
\alias{throw_assert_error}
\title{Throw assert error}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Equivalent of stop(msg) but classed so that other methods can recognise it
}
\usage{throw_assert_error(msg)}
\arguments{
\item{msg}{error message to display}
}

\details{}

\examples{}
\keyword{debugging}
